Keep all comments related types in one module
