+++ packages/liveblocks-core/src/client.ts
@@ -25,13 +25,13 @@ import {
 } from "./room";
 import type { CacheStore } from "./store";
 import { createClientStore } from "./store";
-import type { BaseMetadata } from "./types/BaseMetadata";
+import type { BaseMetadata } from "./protocol/Comments";
 import type { InboxNotificationData } from "./types/InboxNotificationData";
 import type { InboxNotificationDeleteInfo } from "./types/InboxNotificationDeleteInfo";
 import type { OptionalPromise } from "./types/OptionalPromise";
 import type { RoomInfo } from "./types/RoomInfo";
-import type { ThreadData } from "./types/ThreadData";
-import type { ThreadDeleteInfo } from "./types/ThreadDeleteInfo";
+import type { ThreadData } from "./protocol/Comments";
+import type { ThreadDeleteInfo } from "./protocol/Comments";
 
 //
 // Default concrete types for each of the user-provided type placeholders.
