--- packages/liveblocks-core/src/types/BaseMetadata.ts
@@ -1,4 +1,0 @@
-export type BaseMetadata = Record<
-  string,
-  string | boolean | number | undefined
->;
