--- packages/liveblocks-core/src/types/CommentData.ts
@@ -1,37 +1,0 @@
-import type { CommentBody } from "./CommentBody";
-import type { DateToString } from "../lib/DateToString";
-
-export type CommentReaction = {
-  emoji: string;
-  createdAt: Date;
-  users: {
-    id: string;
-  }[];
-};
-
-/**
- * Represents a comment.
- */
-export type CommentData = {
-  type: "comment";
-  id: string;
-  threadId: string;
-  roomId: string;
-  userId: string;
-  createdAt: Date;
-  editedAt?: Date;
-  reactions: CommentReaction[];
-} & (
-  | { body: CommentBody; deletedAt?: never }
-  | { body?: never; deletedAt: Date }
-);
-
-export type CommentDataPlain = Omit<
-  DateToString<CommentData>,
-  "reactions" | "body"
-> & {
-  reactions: DateToString<CommentReaction>[];
-} & (
-    | { body: CommentBody; deletedAt?: never }
-    | { body?: never; deletedAt: string }
-  );
