--- packages/liveblocks-core/src/types/CommentReaction.ts
@@ -1,9 +1,0 @@
-import type { DateToString } from "../lib/DateToString";
-
-export type CommentUserReaction = {
-  emoji: string;
-  createdAt: Date;
-  userId: string;
-};
-
-export type CommentUserReactionPlain = DateToString<CommentUserReaction>;
