--- packages/liveblocks-core/src/types/ThreadDeleteInfo.ts
@@ -1,10 +1,0 @@
-import type { DateToString } from "../lib/DateToString";
-
-export type ThreadDeleteInfo = {
-  type: "deletedThread";
-  id: string;
-  roomId: string;
-  deletedAt: Date;
-};
-
-export type ThreadDeleteInfoPlain = DateToString<ThreadDeleteInfo>;
